fix(auth): address coderabbit review round 2 — extract hook, aria, checkout validation
- Extract useBusinessNameCheck hook (DRY — was duplicated in RegisterStep1 + RegisterSocialStep)
- Add console.error for non-OK responses in business name check
- Add aria-controls to industry combobox for screen reader association
- Add exit-on-failure guards to git checkout commands in CLAUDE.md bootstrap
