+++ frontend/src/components/auth/hooks/index.ts
@@ -5,3 +5,4 @@ export { useReducedMotion } from './useReducedMotion';
 export { useDebounce } from './useDebounce';
 export { useOtpLogic, type UseOtpLogicOptions, type UseOtpLogicReturn } from './useOtpLogic';
 export { useAuthForm, type UseAuthFormOptions, type UseAuthFormReturn } from './useAuthForm';
+export { useBusinessNameCheck } from './useBusinessNameCheck';
